Failed attempt on allow wrapping past 180 degrees (prime meridian) #3.
NoteL when tile wrapping is enabled, leafletjs internally maintains
the longitude as degrees which may be varying odd multiples of 180. To
support this on the PostGIS side, a translation layer would be needed
to make sure the web app and the python code were in agreement about
the frame of reference. This seems more trouble than worth at the
current time.